remove bus_spi.c
